backend/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-filter>=24.2
PyMySQL>=1.1.1
python-dotenv>=1.0.1
Pillow>=10.4
drf-spectacular>=0.27
gunicorn>=23.0

# Tests
pytest>=8.0
pytest-django>=4.8
